Add unregister_and_flush to backend
Adds `unregister_and_flush` to allow a node to unregister from the
backend, but block until all that node's listener callbacks have returned. This is
particularly useful when using async listeners, as they may be running
after the listener has been removed from the backend.

Note that this may also block forever if the listener callbacks are not
designed to return (such as in `mirage-tcpip`).

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>