Add vnetif_stack, monitor function, listen mutex for tests
This commit adds support for creating the V4 stack directly from the
library (Vnetif_stack). Previously we only provided the netif layer, so
the stack had to be assembled manually.

The Vnetif.connect function now accepts an optional monitor
function. This function can be used to capture packets as they are
read/written by the virtual interface. Packets can also be captured by
connecting another interface to the backend and observe traffic from
there, but this may result in a different ordering when the listen
functions are called in async (which they usually are). By capturing on
the interface we're able to see the traffic in the same order as it is
sent/received.

An unlock_on_listen mutex can now be passed to Vnetif.connect. If
specified, the mutex will be unlocked if the listen function is called
on the interface. This is useful for tests where a client thread waits
for a server to listen for new connections. The client can now wait for
the lock to unlock instead of sleeping.

Signed-off-by: Magnus Skjegstad <magnus@skjegstad.com>