Split project in two packages, clean up deps
Vnetif_stack is now bundled in its own package to avoid a circular
dependency on tcpip, which uses regular Vnetif for testing.

Tests are currently only available for the mirage-vnetif-stack package,
as they use tcpip to establish a connection.

Also cleans up the main opam file to remove unused dependencies.